+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.3.9-dev+4.11.2014
 
 - Added the `scrollToLine` and `scrollToRow` methods to `Editor`.
+- Added the `getTokenAt` method to `EditSession` and the `Token` class.
 
 ## 0.3.8+4.11.2014